@@ -14,7 +14,7 @@ description: Demonstrates how to use the international_phone_input plugin.
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.19.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -25,8 +25,7 @@ dependencies:
   cupertino_icons: ^0.1.2
 
   international_phone_input:
-    git:
-      url: git://github.com/niinyarko/flutter-international-phone-input
+    path: ../
 
 dev_dependencies:
   flutter_test:
