@@ -4,12 +4,12 @@ version: 1.0.4
 homepage: https://github.com/niinyarko/flutter-international-phone-input
 
 environment:
-    sdk: ">=2.2.2 <3.0.0"
+    sdk: '>=2.19.0 <3.0.0'
 
 dependencies:
     flutter:
         sdk: flutter
-    libphonenumber: ^1.0.0
+    libphonenumber: ^2.0.2
 
 dev_dependencies:
     test: ^1.6.0
